Ensure accurate fuel amount calculations and data handling
Adjusted fuel amount calculations across client and server to consistently use parseFloat, ensuring proper handling of decimal values from the database. This addresses type mismatches and prevents calculation errors in fuel delivery and order management features.

+++ client/src/pages/ops/orders.tsx
@@ -716,7 +716,7 @@ function OrderCard({ order, position, onAdvanceStatus, getNextStatusLabel, isPen
 
   const handleSave = () => {
     updateOrderMutation.mutate({
-      fuelAmount: parseInt(editedOrder.fuelAmount),
+      fuelAmount: parseFloat(editedOrder.fuelAmount),
       notes: editedOrder.notes,
       address: editedOrder.address,
       city: editedOrder.city,
